fix: Avoid removing `base` from proxied requests to avoid confusion.
The base middleware in Vite will attempt of visits without base when
sending a request that accepts HTML, such as accessing an asset URL
using the browser.

This confuses users, and prevents from easily checking the response in
the browser.

The code was added before Vite had base middleware, so it can be safely
removed now.