fix: Lock dry-cli to 0.6 since 0.7 has breaking changes (close #76)
In dry-cli 0.7, the files logic has been extracted to a separate gem,
dry-files: https://dry-rb.org/news/2021/05/04/introducing-dry-files/

Since several gems still depend on Dry::CLI::Utils::Files, locking the
dependency is the best short term approach. In the future it might be
better to cut the dry-files dependency since it's only used during
installation and the few utilities that are used could be included.