fix: Escaping of asset hosts protocol
Currently if the asset hosts contains a protocol (https://my-cdn.org for example),
the path.join method from node will replace the double forward slash with a single
one. Therefore the final base URL will become something like:

  https:/my-cdn.org/vite

This gives problems if vite has to resolve URLs itself. For example if you have:

  body {
    background-image: url('../background.png')
  }

will resolve in:

  body { background-image: url('https:/my-cdn.org/vite/background.png') }

which in its turn will call this as a relative path.

The fix removes the path.join method with a simple interpolated string.