fix: avoid proxying requests starting with @
Vite is configured to use 'public_output_dir' as the base, which can
be customized by the user in development to not match any of the routes.

Now that most base-related bugs in Vite have been fixed and HMR respects
the base when issuing refresh requests, as well as nested imports, it's
safe to remove this check by default.

If the path starts with that prefix, it will be redirected to Vite.

An exception is when using an empty 'public_output_dir', in which case
it's important to handle virtual requests and still check for `/@`.