feat: expose `ViteRuby.digest` to simplify asset versioning.
One common challenge with single-page apps is refreshing site assets
when they've been changed.

By using a digest of the files watched by Vite Ruby, it's possible to
send it in an HTTP header, allowing the frontend to perform a full page
reload instead of a client-side navigation when the version has changed.

Related: https://inertiajs.com/asset-versioning