Add "main" to package.json to fix eslint issues
Using eslint, I'm getting an `import/no-unresolved: Unable to resolve path to module 'vite-plugin-ruby'` when it clearly exists. Looking at [this ticket](https://github.com/browserify/resolve/issues/222) I believe that is happening because there is no `main` key in the package.json. This updates the package.json to include this.